lab7 peredelal

+++ app/Http/Controllers/ArticleController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Article;
+use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+
 
 class ArticleController extends Controller
 {
@@ -13,7 +16,7 @@ class ArticleController extends Controller
     public function index()
     {
         $articles = Article::latest()->paginate(5);
-        return view('/article/article', ['articles' => $articles]);
+        return view('/article/article', ['articles'=>$articles]);
     }
 
     /**
@@ -21,6 +24,7 @@ class ArticleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Article::class);
         return view('article.create');
     }
 
@@ -29,18 +33,19 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Article::class);
         $request->validate([
             'date' => 'required|date',
-            'title' => 'required|min:5',
-            'text' => 'max:300'
+            'title' => 'required|min:10',
+            'text' => 'max:100'
         ]);
-        $article = new Article();
+        $article = new Article;
         $article->date_public = $request->date;
         $article->title = request('title');
         $article->text = $request->text;
         $article->users_id = auth()->id();
         $article->save();
-        return redirect()->route('article.index')->with('message', 'Create successful');
+        return redirect()->route('article.index')->with('message','Create successful');
     }
 
     /**
@@ -48,8 +53,8 @@ class ArticleController extends Controller
      */
     public function show(Article $article)
     {
-        $article->load('comments');
-        return view('article.show', ['article' => $article]);
+        $comments = Comment::where('article_id', $article->id)->get();
+        return view('article.show', ['article'=>$article, 'comments'=>$comments]);
     }
 
     /**
@@ -57,6 +62,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        Gate::authorize('restore', $article);
         return view('article.edit', ['article'=>$article]);
     }
 
@@ -65,17 +71,18 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        Gate::authorize('update', $article);
         $request->validate([
             'date' => 'required|date',
-            'title' => 'required|min:5',
-            'text' => 'max:300'
+            'title' => 'required|min:10',
+            'text' => 'max:100'
         ]);
         $article->date_public = $request->date;
         $article->title = request('title');
         $article->text = $request->text;
-        $article->users_id = auth()->id();
+        $article->users_id = 1;
         $article->save();
-        return redirect()->route('article.show', ['article'=>$article->id])->with('message', 'Update successful');
+        return redirect()->route('article.show', ['article'=>$article->id])->with('message','Update successful');
     }
 
     /**
@@ -83,7 +90,8 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('delete', $article);
         $article->delete();
-        return redirect()->route('article.index')->with('message', 'Delete Successful');
+        return redirect()->route('article.index')->with('message','Delete successful');
     }
 }
